requirements.txt
Flask
numpy
Pillow
toml
gunicorn
opencv-python
python-dotenv
types-Pillow
